refactor: isolate provider runtime resolution
